+++ src/Hooks/useDrawCut/useDrawCut.ts
@@ -10,6 +10,7 @@ import ValidOp from 'jsts/org/locationtech/jts/operation/valid/IsValidOp';
 
 import * as OlEventConditions from 'ol/events/condition';
 import OlFeature from 'ol/Feature';
+import OlGeometry from 'ol/geom/Geometry';
 import GeometryCollection from 'ol/geom/GeometryCollection';
 import LinearRing from 'ol/geom/LinearRing';
 import LineString from 'ol/geom/LineString';
@@ -18,7 +19,6 @@ import MultiPoint from 'ol/geom/MultiPoint';
 import MultiPolygon from 'ol/geom/MultiPolygon';
 import Point from 'ol/geom/Point';
 import Polygon from 'ol/geom/Polygon';
-import OlGeometry from 'ol/geom/Geometry';
 import OlInteractionDraw, {DrawEvent as OlDrawEvent} from 'ol/interaction/Draw';
 import OlVectorLayer from 'ol/layer/Vector';
 import OlSourceVector from 'ol/source/Vector';